feat: split to jwt/authentication packages
