+++ moit-api/src/main/kotlin/com/mashup/moit/facade/UserFacade.kt
@@ -3,7 +3,7 @@ package com.mashup.moit.facade
 import com.mashup.moit.controller.dto.UserRegisterRequest
 import com.mashup.moit.domain.user.User
 import com.mashup.moit.domain.user.UserService
-import com.mashup.moit.security.getProviderUniqueKey
+import com.mashup.moit.security.authentication.getProviderUniqueKey
 import org.springframework.security.oauth2.core.oidc.user.OidcUser
 import org.springframework.stereotype.Component
