+++ moit-api/src/main/kotlin/com/mashup/moit/security/authentication/JwtAuthentication.kt
@@ -1,4 +1,4 @@
-package com.mashup.moit.security
+package com.mashup.moit.security.authentication
 
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.GrantedAuthority
